DTPRJ-28671,DTPRJ-28673 consent_save_and_exit_options_menu_and_modals
This commit contains implementation of modal window to be able to
agree or disagree with a consent. Also it contains code for
ptions menu with two options "Withdrow Consent" and "Save and Exit".

Overall code was cleaned a little, and overall hierarchy, relations
and integration between objects like Signer, Document and signingService
becomes more intuative and logicall, however not complete.