DTPRJ-28693/feature review_document_before_signing
This PR bring in directive to be able to detect if user scrolled to the bottom
of the documents display view. Also a message was added to the bottom panel
asking to review document before you can start signing.

This PR also contains some cleanups and removal of unused templates.